# API-3.2/openapi-yaml-client-generated/openapi.yaml
openapi: 3.0.0
info:
  title: Shachar and Roman API
  description: API for assigment 3
  contact:
    email: shacharron101@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 3.2.1-oas3
servers:
- url: http://localhost:4000
...
paths:
  /user/register:
    post:
      tags:
      - user
      summary: Add new user
      description: This will add a new user to the site
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          description: User created
        "400":
          description: The request is invalid, There is empty fields.
        "405":
          description: E-Mail already exist
        "409":
          description: Username already exist
  /user/login:
    post:
      tags:
      - user
      summary: login
      description: Logs user into the system
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        "200":
          description: login succeeded
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: session=87s97sdhgf279293f8w0e9
        "401":
          description: Username or Password incorrect.
  /user/logout:
    get:
      tags:
      - user
      description: logout for user, we delete the cookie
      responses:
        "200":
          description: logout succeeded.
        "401":
          description: error
      security:
      - cookieAuth: []
...
components:
  schemas:
    User:
      required:
      - country
      - email
      - firstname
      - imageUrl
      - lastname
      - password
      - username
      type: object
      properties:
        username:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        country:
          type: string
        email:
          type: string
        password:
          type: string
        imageUrl:
          type: string
        recipes:
          type: array
          items:
            $ref: '#/components/schemas/Recipe'
        family_recipes:
          type: array
          items:
            $ref: '#/components/schemas/Recipe'
        user_last_seen:
          type: array
          items:
            type: integer
        user_favorites:
          type: array
          items:
            type: integer
        last_serach:
          type: string
      description: User object
      example:
        username: Shacharron101
        firstName: Shachar
        lastName: Ron
        country: Israel
        email: shacharron101@gmail.com
        password: qwe123
        imageUrl: https://cloudinary.com/
        last_serach: apple
...
    Recipe:
      required:
      - author
      - instructions
      - rec_gluten
      - rec_image_url
      - rec_ingredients
      - rec_name
      - rec_number_of_dishes
      - rec_popularity
      - rec_time
      - rec_vegan
      - rec_vegetarian
      type: object
      properties:
        rec_id:
          type: integer
        author:
          type: integer
        rec_name:
          type: string
        rec_image_url:
          type: string
        rec_time:
          type: integer
        rec_popularity:
          type: integer
        rec_vegan:
          type: boolean
        rec_vegetarian:
          type: boolean
        rec_gluten:
          type: boolean
        rec_ingredients:
          type: string
        instructions:
          type: string
        rec_number_of_dishes:
          type: integer
      example:
        rec_name: Pizza
        rec_time: 10
        rec_popularity: 10
        rec_vegan: false
        rec_vegetarian: false
        rec_gluten: false
        rec_ingredients:
        - 1. dough
        - 2. cheese
        - 3. tomatos
        - 4. olive
        instructions: step1- Make the dough step2- Put on the dough the tomatoes,
          cheese and olives. step3- Put the pizza in the oven to 180 degrees for 20
          minutes.
        rec_number_of_dishes: 2
    body:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: Contain only letters
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: Contains letters with at least 1 numerical character and 1
            special characters
      example:
        username: roman
        password: "1234"
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: session
      in: cookie